Tests/LibWeb/Text/expected/HTML/script-blocking-style-sheets-do-not-block-child-documents.txt

classic script executed while parent style sheet was pending: true
module script executed while parent style sheet was pending: true
